Refactor skill value types to double, UI/JSON improvements
- Changed BlackboardEntry::value and Skill::duration from float to double for higher precision.
- Updated JSON loading to use "version" instead of "dataVersion".
- Applied snapping to duration and effect values for consistency.
- Improved ImGui effect value input with finer steps and formatting.
- Cleaned up redundant mouse click handling in range grid editor.
- Updated imgui.ini with new window positions, sizes, and added "Skill Edit" window entry.

+++ AKDataEditor/Skill.h
@@ -6,7 +6,7 @@
 struct BlackboardEntry
 {
 	std::string key;
-	float value;
+	double value;
 	NLOHMANN_DEFINE_TYPE_INTRUSIVE(BlackboardEntry, key, value)
 };
 
@@ -33,7 +33,7 @@ struct Skill
 	std::string name;
 	std::string description;
 	int skillType;			// 0. passive, 1. manual, 2. auto
-	float duration;
+	double duration;
 	SpData spData;
 	std::vector<SkillRange> range;
 	std::vector<BlackboardEntry> blackboard;
